requirements.txt
requests
python-dotenv
langchain-mcp-adapters
langchain-core
Flask[async]
PyYAML
langchain